backend/requirements.txt
Django==2.2.8
psycopg2-binary==2.8.4
gunicorn==20.0.4
djangorestframework==3.11.0
django-templated-mail==1.1.1
Pillow==6.2.2
python-magic==0.4.15

djoser==2.0.3
django-cors-headers==3.2.0
django-oauth-toolkit==1.3.2

django-environ==0.4.5

elasticsearch>==7.0.0,<8.0.0
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.5

django-elasticsearch-debug-toolbar==2.0.0
django-debug-toolbar-force==0.1.7
django-debug-toolbar==2.1

mixer==6.1.3
Faker==3.0.0

django-request-logging==0.7.0
requests==2.23.0

pylint==2.4.4
flake8==3.7.9
pytest==5.2.2
pytest-django==3.7.0
pytest-cov==2.8.1
pycodestyle==2.5.0
autopep8==1.5.0
black==19.10b0